folders redesign. backlinks still bad
