+++ app/core/__init__.py
@@ -3,8 +3,13 @@ from .exceptions import (
     CommonException,
     DoNotUsuRefreshToken,
     InternalServerError,
+    IssueNotFoundError,
     NoRefreshToken,
     NotFoundException,
+    SprintNotFoundError,
+    UserNotFoundError,
+    ValidationError,
+    WorkplaceNotFoundError,
 )
 from .mongo_session import MongoManager
 from .redis_session import Redis
@@ -18,4 +23,9 @@ __all__ = [
     "DoNotUsuRefreshToken",
     "Redis",
     "MongoManager",
+    "IssueNotFoundError",
+    "SprintNotFoundError",
+    "WorkplaceNotFoundError",
+    "ValidationError",
+    "UserNotFoundError",
 ]
