--- app/routers/auth/__init__.py
@@ -1,3 +1,0 @@
-from .schemas import User, UserAssignedWorkplace
-
-__all__ = ["User", "UserAssignedWorkplace"]
