--- app/routers/issue/__init__.py
@@ -1,3 +1,0 @@
-from .schemas import Issue
-
-__all__ = ["Issue"]
