--- app/routers/sprint/__init__.py
@@ -1,3 +1,0 @@
-from .schemas import Sprint
-
-__all__ = ["Sprint"]
