--- app/routers/workplace/__init__.py
@@ -1,3 +1,0 @@
-from .schemas import Workplace
-
-__all__ = ["Workplace"]
