+++ app/schemas/sprint.py
@@ -6,7 +6,10 @@ from beanie import Document, Link
 from pydantic import BaseModel, Field, model_validator
 
 from app.core.exceptions import ValidationError
-from app.routers.issue import Issue
+
+from .issue import Issue
+
+# from .workplace import Workplace
 
 
 class SprintCreation(BaseModel):
@@ -22,6 +25,7 @@ class SprintCreation(BaseModel):
 
 
 class Sprint(Document, SprintCreation):
+    # workplace_id: BackLink[Workplace]
     workplace_id: UUID
     id: UUID = Field(default_factory=uuid4)
     issues: List[Link[Issue]] = Field(default_factory=list)
